@@ -16,7 +16,7 @@ class CaptchasTableTest extends TestCase {
 	 *
 	 * @var array<string>
 	 */
-	protected $fixtures = [
+	protected array $fixtures = [
 		'plugin.Captcha.Captchas',
 	];
 
